+++ services/http_server.js
@@ -1,8 +1,8 @@
 module.exports = async function(container, app, config) {
   let httpApp = await container.create(require('../lib/http_server/server'))
 
-  const isRepl = app.getAppCommand() === 'repl'
+  const isEnvWithoutServer = app.isREPL || app.isTask
   const forceListen = config.get('quadro.http.force')
 
-  if (!isRepl || forceListen) return httpApp.listen()
+  if (!isEnvWithoutServer || forceListen) return httpApp.listen()
 }